Added support for lights, xfan, sleep and powersave mode. (#16)
While trying to understand the code and blocks used in the component I made some changes to # remarks and _LOGGER.info statements to enhance readability.
One of the pro's of the mqtt version was the possibility to add mqtt switches used for setting severall options in de airco like xfan, lights etc. So...tried to find a way to implement those in this custom component and it seems to me I have a working/bug-free version here.

For the options to work you need to create input_booleans and pass them to the component.

Can you check it out and let me know what you think of this update?

In summary I changed the following stuff:
- Added support for passing input_booleans to the component to set options for lights, xfan, sleep and powersave modes. 
- Removed unnecessary code (mostly default configs, i.m.h.o not needed due to syncing with hvac device on startup, some of them where not used in the code).
- Corrected severall code imperfections (timeout variable was not used in the corresponding code, use of DEFAULT_... instead of self._... in code etc).
- Improved consistancy in list names.
- Changed some of the _logger.info for more readability while learning the code and component.
- Changed the handling of recent implemented "off" operations mode (avoid setting unsupported "mod" numbers to the HVAC, and keeping the current operation mode in HA in sync with the HVAC).

Extra info:
xfan can be used in cool and dry mode.
when xfan is turned on the airco fan will run for a while when cooling is switched off to allow the unit to "dry" before it switches off and closed to avoid unpleasent smell.
powersave can only be used in cool mode.
Gree documentation is not very informing what it accualy does but it seems the inside temperature will vary a bit depending on the outside temperature. I'll try to find more info on this one.
sleep mode can only be used while cooling and is a special mode for sleeping. temperature will be lower when unit is switched on and than increasing a bit over time (when you - hopefully - are in deeper sleep). My unit has several sleep modes. Not sure yet how to set the different modes (SwhSlp only seems to be either 0 or 1).