Device availability (#163)
Solved problem of device unavailability if home assistant starts when devices are offline or have no power. From now on, devices are correctly added to the home assistant. There is a specific time interval when the home assistant checks whether the devices are online or not. If not, device is marked as unavailable until it appears on the network.